Move TranslationBride to a separate file
Add spdx license headers in all files too
