+++ Main.qml
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: GPL-3.0-only
 import QtQuick
 import QtCore
 import QtQuick.Controls
