+++ config.hpp.in
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: GPL-3.0-only
 #ifndef CONFIG_HPP
 #define CONFIG_HPP
